create project.loader
